Update oauth.php
Added REDIRECT_URI to the $permission_url which is now required by Shopify

"{redirect_uri} - (Required) substitute this with the URL where you want to redirect the users after they authorize the client. The complete URL specified here must be identical to one of the Application Redirect URLs. Note: in older applications, this parameter was optional, and redirected to the Application Callback URL when no other value was specified."  - https://docs.shopify.com/api/authentication/oauth#asking-for-permission